Changed vertex index size to 32 bits

+++ engine/include/IndexBuffer.hpp
@@ -7,9 +7,9 @@ using std::vector;
 
 class IndexBuffer : public GraphicsBuffer {
 public:
-	IndexBuffer(shared_ptr<const GraphicsContext> context, const vector<uint16_t>& indices);
+	IndexBuffer(shared_ptr<const GraphicsContext> context, const vector<uint32_t>& indices);
 
 
 private:
-	void fillMemory(const vector<uint16_t>& indices) const;
+	void fillMemory(const vector<uint32_t>& indices) const;
 };
